Flatten nesting of test backends and rename the folder appropriately.
