+++ backends/p4tools/modules/testgen/targets/bmv2/test_backend/protobuf.cpp
@@ -1,4 +1,4 @@
-#include "backends/p4tools/modules/testgen/targets/bmv2/backend/protobuf/protobuf.h"
+#include "backends/p4tools/modules/testgen/targets/bmv2/test_backend/protobuf.h"
 
 #include <algorithm>
 #include <filesystem>
